Edited Unit Titles
I edited this to name the units so people just have a good idea of how things lump together and it feels like "real school" I'm assuming the checks for understanding are built into the Kat costume website and bakery. We may need an easier one for the very first?